+++ src/pages/PostEditPage/DropdownMenu/index.tsx
@@ -6,6 +6,9 @@ import {
   MenuForm,
 } from './DropdownMenuStyled';
 import { useEffect, useState } from 'react';
+import { useSelector } from 'react-redux';
+import { getChannel } from '@/slices/channel';
+import { useDispatch, RootState } from '@/store';
 
 export interface Channel {
   _id: string;
@@ -18,6 +21,13 @@ interface DropdownMenuProps {
 }
 
 const DropdownMenu = ({ channelId, setChannelId }: DropdownMenuProps) => {
+  const dispatch = useDispatch();
+
+  useEffect(() => {
+    dispatch(getChannel());
+  }, [dispatch]);
+
+  const channels = useSelector((state: RootState) => state.channel.channels);
   const [isMenuOpen, setIsMenuOpen] = useState(false);
   const [channelTitle, setChannelTitle] = useState('채널 선택');
 
@@ -31,9 +41,9 @@ const DropdownMenu = ({ channelId, setChannelId }: DropdownMenuProps) => {
   };
 
   useEffect(() => {
-    const selectedChannel = itemListData.find((item) => item._id === channelId);
+    const selectedChannel = channels.find((item) => item._id === channelId);
     setChannelTitle(selectedChannel ? selectedChannel.name : '채널 선택');
-  }, [channelId]);
+  }, [channelId, channels]);
 
   return (
     <MenuForm>
@@ -47,7 +57,7 @@ const DropdownMenu = ({ channelId, setChannelId }: DropdownMenuProps) => {
       </DropdownButton>
       {isMenuOpen && (
         <DropdownContent>
-          {itemListData.map((item) => (
+          {channels.map((item) => (
             <ListItemButton
               type='button'
               key={item._id}
@@ -62,37 +72,3 @@ const DropdownMenu = ({ channelId, setChannelId }: DropdownMenuProps) => {
 };
 
 export default DropdownMenu;
-
-//임시 더미 데이터
-const itemListData = [
-  {
-    authRequired: false,
-    posts: [],
-    _id: '6587c05d83003970282b863e',
-    name: '연예',
-    description: 'Test',
-    createdAt: '2023-12-24T05:23:41.475Z',
-    updatedAt: '2023-12-24T05:23:41.475Z',
-    __v: 0,
-  },
-  {
-    authRequired: false,
-    posts: [],
-    _id: '6587s',
-    name: '스포츠',
-    description: 'Test',
-    createdAt: '2023-12-24T05:23:41.475Z',
-    updatedAt: '2023-12-24T05:23:41.475Z',
-    __v: 0,
-  },
-  {
-    authRequired: false,
-    posts: [],
-    _id: '6587c',
-    name: '잡담',
-    description: 'Test',
-    createdAt: '2023-12-24T05:23:41.475Z',
-    updatedAt: '2023-12-24T05:23:41.475Z',
-    __v: 0,
-  },
-];
